remove ps2 naming, go to usb
