+++ CMakeLists.txt
@@ -16,7 +16,7 @@ project(pico-ps2kb C CXX ASM)
 
 pico_sdk_init()
 
-add_subdirectory(ps2famikb-lib)
+add_subdirectory(usb2famikb-lib)
 add_subdirectory(pico-pio-usb)
 
 add_executable(pico-ps2kb)
@@ -32,7 +32,7 @@ target_link_libraries(pico-ps2kb PUBLIC
     hardware_i2c
     pico_stdlib
     pico_multicore
-    ps2famikb-lib
+    usb2famikb-lib
     
 )
